requirements.txt
click>=8.1.7
requests>=2.31.0
tqdm>=4.66.4
Pillow>=10.3.0
geopy>=2.4.1
tenacity>=8.2.3